test(evals): lift graph_store fixture to shared conftest
Both eval files defined an identical-shape graph_store fixture
(temp SQLite path, try/finally close()). Move it once into
evals/conftest.py and drop the local copies. Rename the merge
file's `store` parameter to `graph_store` for cross-file symmetry,
and remove the now-unused GraphMemoryStore imports plus the
`store = graph_store` rebinding in the recency suite.
